adjust minor changes in getting total amount in order model

+++ database/migrations/2025_06_26_124844_create_orders_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('reference_id');
             $table->foreignId('user_id')->constrained()->onDelete('cascade');
-            $table->unsignedInteger('total_amount');
+            $table->decimal('total_amount', 10, 2);
             $table->enum('type', ['government', 'walk_in', 'project_based', 'online']);
             $table->enum('status', ['pending', 'completed', 'expired'])->default('pending');
             $table->date('expiry_date')->nullable();
